requirements.txt
streamlit
openai
python-dotenv
requests
pandas
langchain
langsmith
streamlit-folium